requirements.txt
fastapi==0.116.1
folium==0.20.0
geopandas==1.1.1
geopy==2.4.1
openrouteservice==2.3.3
pandas==2.3.1
pip==25.1.1
python-dotenv==1.1.1
requests==2.32.4
uvicorn==0.35.0